pkgdepR: Statically Determine Function Dependencies Between Packages

Statically determine and visualize the function dependencies within and across packages. This may be useful for managing function dependencies across a code base of multiple R packages.

Details

The only function you're likely to need from pkgdepR is deps, which returns an object of class pkgdepR.

An object of class pkgdepR is a list with three named objects:

funs

a data frame describing the functions. Contains columns id, label, package, exported, group, and name.

links

a data frame containing the linkages between functions. Contains columns from and to.

pkg

a character vector containing the packages explored.

See also

Author

Maintainer: Ed Peyton edppeyton@gmail.com (ORCID)