Visualizing package dependencies

This function is a simple wrapper for plotting a network visualization using visNetwork.

Usage

# S3 method for class 'pkgdepR'
plot(
  x,
  width = NULL,
  height = NULL,
  main = NULL,
  submain = NULL,
  alpha = 0.8,
  footer = NULL,
  background = "rgba(0, 0, 0, 0)",
  n,
  m,
  ...
)

Arguments

x

An object of class pkgdepR.

width

The width of the vis.js render.

height

The height of the vis.js render.

main

The title. To remove the title, pass list(text = NULL).

submain

The subtitle. To remove the subtitle, pass list(text = NULL).

alpha

A transparency value to use for colors. Must be between 0 and 1.

footer

A character or a named list. See visNetwork.

background

A background color. See visNetwork.

n

(Optional) The number of colours to request from viridis. Allows the user to set a more granular palette.

m

(Optional) The subset of colours of the custom palette (specified by n) to use in the plot.

...

Other arguments passed onto viridis.

Value

An object of classes visNetwork and htmlwidget.

Examples

library(pkgdepR)
deps(pkg = "pkgdepR") %>% plot(option = "E", direction = -1)

deps(pkg = "pkgdepR") %>% plot(option = "D", alpha = 0.5, main = list(text = NULL))

deps(pkg = "pkgdepR") %>% 
  plot(option = "D", alpha = 0.8, main = list(text = NULL)) %>% 
  visNetwork::visInteraction(dragNodes = FALSE)